Make properly shutdown server after tests, bug fixes
