+++ .gitignore
@@ -19,3 +19,6 @@ __pycache__/
 env/
 venv/
 ENV/
+
+# istanbul
+.nyc_output/
